requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.8.0

# Training
pandas>=2.0.0
scikit-learn>=1.3.0
mlflow>=2.8.0
tqdm>=4.65.0

# API and UI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
requests>=2.31.0

# Config
PyYAML>=6.0.1

# Evaluation / visualization (train pipeline)
seaborn>=0.13.0

# Tests
pytest>=7.4.0
